Refactor installed/current game management into new GamesManager class
Inherit LootState from it for now. GamesManager is an abstract class
that leaves finding game paths and initialising Game objects up to
the implementor, so it's simpler to test.

This also changes LootState::storeGameSettings()'s behaviour, it now
only updates the current game's Game object, and replaces the others.
The current object is also replaced if the game's path, local path
or master have changed, as these invalidate the current game handle.